Ajuste nos endpoints

+++ services/autor.service.js
@@ -1,4 +1,6 @@
 import AutorRepository from "../repositories/autor.repository.js";
+import LivroRepository from "../repositories/livro.repository.js";
+
 
 async function createAutor(autor) {
   return await AutorRepository.insertAutor(autor);
@@ -13,6 +15,10 @@ async function getAutor(id) {
 }
 
 async function deleteAutor(id) {
+  const livros = await LivroRepository.getLivrosByAutor(id);
+  if (livros.length > 0){
+    throw new Error ("Existe livros cadastrados para esse Autor, não possível excluir");
+  }
    await AutorRepository.deleteAutor(id);
 }
